% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Returns a data frame representation of various kinds of Facebook collections}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}

\item{row.names}{If set to \code{TRUE}, names the rows of the returned data frame with IDs of the elements}

\item{optional}{Not used in this context.}
}
\description{
Returns a data frame representation of various kinds of Facebook collections
}
\keyword{internal}

