% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\name{as.list.FacebookGenericCollection}
\alias{as.list.FacebookGenericCollection}
\title{Returns a list from a Facebook collection}
\usage{
\method{as.list}{FacebookGenericCollection}(x, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}
}
\description{
This generic return a valid list representation of various Facebook collections
}
\keyword{internal}
