% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.advanced.R
\docType{package}
\name{facebook.advanced}
\alias{facebook.advanced}
\alias{facebook.advanced-package}
\title{Access to Facebook API v2 via R}
\description{
This package provides a series of functions that allow R users
to access Facebook's API to get information about users and posts,
and collect public status updates that mention specific keywords.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{FacebookUsersCollection}},
\code{\link{FacebookPagesCollection}}, \code{\link{FacebookPostsCollection}},
\code{\link{FacebookCommentsCollection}}
}
\keyword{facebook}
\keyword{package}

