% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/formatFbDate.R
\name{formatFbDate}
\alias{formatFbDate}
\title{Convert a Facebook datetime string into a valid R datetime}
\usage{
formatFbDate(datestring, format = "datetime")
}
\arguments{
\item{datestring}{The date to convert in the character format used by Facebook.}

\item{format}{Set to \code{"datetime"} to convert into a POSIX value or \code{"date"}
to convert into a Date. Any other values will make the function returns \code{NULL}.}
}
\description{
\code{facebook.formatFbDate} converts the string datetime returned by Facebook in a valid POSIX format.
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

