% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.object.likes.R
\name{facebook.object.likes}
\alias{facebook.object.likes}
\title{Pull all the public likes of Facebook users or pages}
\usage{
facebook.object.likes(id, token = NULL, parameters = list(),
  n = getOption("facebook.maxitems"), .progress = create_progress_bar())
}
\arguments{
\item{id}{An existing \code{\link{FacebookUsersCollection}} or \code{\link{FacebookPagesCollection}}}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If \code{NULL} and \code{id} is a Collection, get that one instead. Otherwise, no query is performed
to the Facebook Graph API and an empty Collection is returned}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}}

\item{n}{An integer value with the maximum number of participants to be pulled for each conversation in \code{id}. It can be set to \code{Inf}
to pull out any participant of a given conversation and assumes the default value to \code{facebook.maxitems} global option if missing.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console.}
}
\value{
A collection of mixed likes in a \code{\link{FacebookMixedCollection-class}} object with the \code{id} and the \code{type} for
each element included.
}
\description{
\code{facebook.object.likes} pulls information about the likes from a list of Facebook IDs and/or names of users or pages and push them into a
\code{\link{FacebookMixedCollection-class}} instance.
}
\details{
This function requires the use of a OAuth token with \code{user_likes}
permission granted. After the introduction of version 2.0 of the Graph API,
only likes from users who are using the application that you used to generate the
token to query the API will be returned.

Only the \code{id} and the \code{type} is returned in a mixed collection.
Then, a proper collection for each type must be built accordingly.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")

## Returns the id and the type of the current user's likes
 me.likes <- facebook.users.likes(id = "me", token = fb_oauth)

## Do the same, but starting from a users collection
 me.likes <- FacebookUsersCollection(id="me", fb_oauth) \%>\% facebook.users.likes()

## Build a pages collection from all the pages (and ONLY the pages)
## the current user likes
 me.likes.pages <- me.likes \%>\% FacebookPagesCollection()
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.search}}, \code{\link{fbOAuth}}
}

