% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.pages.conversations.R
\name{facebook.posts}
\alias{facebook.posts}
\title{Extract information about a set of public Facebook posts}
\usage{
facebook.posts(posts, token,
  fields = "id,from.fields(id,name),message,created_time,type,link,name",
  .progress = create_progress_bar())
}
\arguments{
\item{posts}{A vector or a comma-delimited string of post IDs}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{fields}{vector or comma-delimited string with the post-level details to get.
\code{n}.}
}
\description{
\code{facebook.posts} retrieves information about a set of public Facebook posts.
}
\details{
\code{facebook.posts} returns data about posts but doesn't return lists of comments or likes (altough it will return a summary view of both).
You can actually get some informations about comments and likes using fields nesting (see below), but this is not actually
recommended.

Consider using the twin functions \code{\link{facebook.comments}}, \code{\link{facebook.likes}} to focuse of these nodes
of data.

Due to the network-graph nature of Facebook data model,
you can always specify fields details for each field eventually nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{name} for the \code{from} node, this clause is valid among others:
\code{from.fields(id,name)}.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about Facebook's Facebook Page
	load("fb_oauth")
	fb.page <- getPages("9thcirclegames", token=fb_oauth)
## Getting information about 10 most recent posts
	post.details <- facebook.posts(post=fb.page$posts[10,]$id, token=fb_oauth)
}
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{facebook.pages}}, \code{\link{facebook.comments}}, \code{\link{facebook.likes}}, \code{\link{fbOAuth}}
}

