% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.S4.R
\docType{package}
\name{facebook.S4}
\alias{facebook.S4}
\alias{facebook.S4-package}
\title{Access to Facebook API v2 via R}
\description{
This package provides a series of functions that allow R users
to access Facebook's API and builds collectiond of items
that model and mimic the graph organization and network relations
between different elements.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{FacebookUsersCollection}},
\code{\link{FacebookPagesCollection}}, \code{\link{FacebookPostsCollection}},
\code{\link{FacebookCommentsCollection}}
}
\keyword{facebook}
\keyword{package}

