% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookAlbumsCollection.R
\name{FacebookAlbumsCollection}
\alias{FacebookAlbumsCollection}
\title{Build a collection of Facebook albums}
\usage{
FacebookAlbumsCollection(id, token = NULL, parameters = list(),
  fields = c("id", "cover_photo", "from.fields(id,name)", "created_time",
  "description", "name", "count"), n = getOption("facebook.maxitems"),
  metadata = FALSE, .progress = create_progress_bar())
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any
of the supported types (see below).}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If token \code{NULL} and \code{id} is a collection,
use the token of the source collection. Otherwise, no query is performed
to the Facebook Graph API and an empty collection is returned.}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}.}

\item{fields}{A character vector with the fields to get for each \code{id}. If no value
for a given field is found, it will be set to \code{NULL}.}

\item{n}{If \code{id} is an iterable collection, then \code{n} is the maximum number of albums to be pulled for each element of the source collection
in \code{id}. It can be set to \code{Inf} to pull out any available public album and assumes the default value from the value
of \code{facebook.maxitems} global option if missing. If \code{id} is not a collection or cannot be iterated, the parameter is ignored.}

\item{metadata}{If set to \code{TRUE}, the medatata for each ID is pulled with the data
and the \code{type} slot is fed accordingly.
Please note that setting this to \code{TRUE} could considerably
slow down the execution time, as more queries are needed.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console. See \link[plyr]{create_progress_bar} for details.}
}
\value{
A collection of albums in a \code{\link{FacebookAlbumsCollection-class}} object.
}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook albums and build a \code{\link{FacebookAlbumsCollection-class}}
instance.
}
\details{
\code{FacebookAlbumsCollection} is the constructor for the \code{\link{FacebookAlbumsCollection-class}}.
It returns data about albums but doesn't return lists of comments or likes, altough it \strong{does} return
an approximate total count for both (depending on privacy settings of the users).
}
\section{Nesting fields}{

Due to the network-graph nature of Facebook data model,
you can specify fields details for each field nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{source} for the \code{cover} field, this is valid among others:
\code{cover.fields(id,source)}.

Following the same philosophy, if you need only \code{id} and \code{name} for the \code{from} node
you can use \code{from.fields(id,name)}.
}

\section{Valid sources}{

Instead of a character vector, one of these collections can also be passed as parameter in \code{id}:
\itemize{
 \item{\code{\link{FacebookPagesCollection-class}} will build a collection with
 the albums created on the pages in the source collection.}
 \item{\code{\link{FacebookUsersCollection-class}} will build a collection with
 the albums created the users in the source collection. Users must have granted
 the \code{user_photos} permission to the current app to be able to perform this action.}
}
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")

## Getting information about two example Facebook Pages
 fb.pages <- FacebookPagesCollection(id = c("9thcirclegames",
                                           "NathanNeverSergioBonelliEditore"),
                                     token = fb_oauth)

## Pull at most 10 albums from each page
 fb.albums <- FacebookAlbumscollection(id = fb.pages, token = fb_oauth, n = 10)

## Pull all the available albums from each page
 fb.albums.inf <- FacebookAlbumscollection(id = fb.pages, token = fb_oauth, n = Inf)

## Convert the collection to a data frame
fb.albums.df <- as.data.frame(fb.albums)

# The same as before in a more compact fashion using the pipe operator
# chaining from a Pages Collection
 fb.albums.pipe <- FacebookPagesCollection(id = c("9thcirclegames",
                                                "NathanNeverSergioBonelliEditore"),
                                          token = fb_oauth) \%>\% FacebookAlbumsCollection(n = 10)
}
}
\author{
Gabriele Baldassarre \url{https://gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookPagesCollection}}, \code{\link{FacebookUsersCollection}}, \code{\link{fbOAuth}}

Other Facebook Collection Costructors: \code{\link{FacebookCommentsCollection}};
  \code{\link{FacebookConversationsCollection}};
  \code{\link{FacebookLikesCollection}};
  \code{\link{FacebookMessagesCollection}};
  \code{\link{FacebookPagesCollection}};
  \code{\link{FacebookPhotosCollection}};
  \code{\link{FacebookPostsCollection}};
  \code{\link{FacebookUsersCollection}}
}

