% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookConversationsCollection.R
\name{FacebookConversationsCollection}
\alias{FacebookConversationsCollection}
\title{Build a Collection of Facebook Inbox Conversations}
\usage{
FacebookConversationsCollection(id, token = NULL, parameters = list(),
  fields = c("id", "snippet", "updated_time", "message_count"), feed = TRUE,
  n = getOption("facebook.maxitems"), metadata = FALSE,
  .progress = create_progress_bar())
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any of the supported types}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If \code{NULL} and \code{id} is a Collection, get that one instead. Otherwise, no query is performed
to the Facebook Graph API and an empty Collection is returned}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}}

\item{fields}{A character vector or a comma-delimited string with the list of fields to get for each ID. If no value for a given fields is found, it will
be set to \code{NULL}}

\item{n}{If \code{id} is a Collection, then \code{n} is the maximum number of conversations to be pulled for any element of the Collection in \code{id}.
Otherwise, the parameter is ignored. It can be set to \code{Inf} to pull out any available public post and assumes the default value from the value
of \code{facebook.maxitems} global option if missing.}

\item{metadata}{If set to \code{TRUE}, the medatata from each ID is pulled with the data and the \code{type} slot is fed. Please note
that setting this to \code{TRUE} could considerably slow down the execution time, as more queries are needed.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console.}
}
\value{
A collection of posts in a \code{\link{FacebookPostsCollection-class}} object.
}
\description{
Connect to Facebook Graph API, get information from a list of inbox conversations and build a \code{\link{FacebooConversationsCollection-class}}
instance.  Then, a \code{\link{FacebookMessagesCollection-class}} can be built to pull out the message of each conversation.
}
\details{
\code{FacebookConversationssCollection} is the constructor for the \code{\link{FacebookConversationssCollection-class}}.
It returns data about inbox conversations of users or pages but doesn't return lists of messages or users involved
(altough it will return a summary view of both).

Consider using the twin functions \code{\link{FacebookMessagessCollection}}, \code{\link{FacebookUserssCollection}} to focus on these kinds
of items.

Use a \code{\link{FacebookPagesCollection-class}} as \code{id} to get a list of inbox conversations
related to a page (a page access token with \code{read_page_mailboxes} is needed) or a \code{\link{FacebookUserssCollection-class}}
(a user access token with \code{read_mailbox} is needed and \strong{the user must be a developer for the app making the request}).

Due to the network-graph nature of Facebook data model,
you can always specify fields details for each field eventually nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{name} for the \code{from} node, this clause is valid among others:
\code{from.fields(id,name)}.

Assuming the senders granted the \code{user_posts} permission in the scope of the current application, you can pass
a collection of \code{FacebookPostsCollection} as \code{id} parameter. In this cases, a collection of sharedposts from the
given posts will be fed.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")

## Getting information about two example Facebook Pages
 fb.pages <- FacebookPagesCollection(id = c("9thcirclegames", "NathanNeverSergioBonelliEditore"), token = fb_oauth)

}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookPagesCollection}}, \code{\link{FacebookUsersCollection}}, \code{\link{fbOAuth}}

Other Facebook Collection Costructors: \code{\link{FacebookCommentsCollection}};
  \code{\link{FacebookLikesCollection}};
  \code{\link{FacebookPostsCollection}}
}

