% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-class.R
\docType{class}
\name{FacebookPagesCollection-class}
\alias{FacebookPagesCollection-class}
\title{Class for representing a Collection of Facebook pages}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook pages and build a \code{FacebookPagesCollection-class}
instance.

Use the \code{\link{FacebookPagesCollection}} constructor for building a instance of this class, as it provides better
sanity and validation checks.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

