% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookUsersCollection-class.R
\docType{class}
\name{FacebookUsersCollection-class}
\alias{FacebookUsersCollection-class}
\title{Class for representing a collection of Facebook users}
\description{
This class is used to collect Facebook pages in a convenient structure.

Use the \code{\link{FacebookUsersCollection}} constructor for building a instance of this class, as it provides better
sanity and validation checks.
}
\author{
Gabriele Baldassarre \url{https://gabrielebaldassarre.com}
}
