% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\docType{methods}
\name{c,FacebookGenericCollection-method}
\alias{c,FacebookGenericCollection-method}
\title{Combine two or more Facebook collections}
\usage{
\S4method{c}{FacebookGenericCollection}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{the first collection to chain into}

\item{...}{the other collections to chain from}

\item{recursive}{not used in this context}
}
\description{
This method combines two or more Facebook collections of the same kind. Please note that duplicates are removed unless they have
different parents.
}
