% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\docType{methods}
\name{as.character,FacebookGenericCollection-method}
\alias{as.character,FacebookGenericCollection-method}
\title{Return a comma-separated string of all the IDs of the given Facebook collection}
\usage{
\S4method{as.character}{FacebookGenericCollection}(x)
}
\arguments{
\item{x}{a valid not-null Facebook Collection}
}
\description{
This method serializes a Facebook collection extracting all the IDs of its member into a
comma-separated string. This could be useful for lazy chaining into other Facebook queries
or for debugging purposes.
}
