% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\docType{methods}
\name{length,FacebookGenericCollection-method}
\alias{length,FacebookGenericCollection-method}
\alias{length,ANY-method}
\title{Return the number of items in a given Facebook collection}
\usage{
\S4method{length}{FacebookGenericCollection}(x)
}
\arguments{
\item{x}{a valid not-null Facebook colle}
}
\description{
This method returns the number of members in a given Facebook collection.
}
