% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookCommentsCollection-methods.R
\name{is.CommentsCollection}
\alias{is.CommentsCollection}
\title{Checks if the argument is a valid collection of comments}
\usage{
is.CommentsCollection(x)
}
\description{
\code{is.CommentsCollection} checks if \code{x} is a valid collection of Facebook Comments built using \code{\link{FacebookCommentsCollection}}.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

