% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookLikesCollection-methods.R
\name{is.LikesCollection}
\alias{is.LikesCollection}
\title{Checks if the argument is a valid collection of likes}
\usage{
is.LikesCollection(x)
}
\description{
\code{is.CommentsCollection} checks if \code{x} is a valid collection of Facebook Likes built using \code{\link{FacebookLikesCollection}}.
}
\details{
Although they are named the same in the Facebook jargon, posts/comments likes and users'likes are different entitites.
This method covers the first, while \code{\link{isUsersLikesCollection}} the latter.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

