% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookUsersCollection-methods.R
\name{is.UsersCollection}
\alias{is.UsersCollection}
\title{Checks if the argument is a valid Collection of users}
\usage{
is.UsersCollection(x)
}
\description{
\code{is.UsersCollection} checks if \code{x} is a valid collection of Facebook users built using \code{\link{FacebookUsersCollection}}.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

