% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\docType{methods}
\name{[,FacebookGenericCollection-method}
\alias{[,FacebookGenericCollection-method}
\title{Return parts of a Facebook collection}
\usage{
\S4method{[}{FacebookGenericCollection}(x, i, j, drop)
}
\arguments{
\item{x}{A valid collection of Facebook elements}

\item{i}{slicing on the first dimension index}

\item{j}{not used in this context}

\item{drop}{not used in this context}

\item{...}{not used in this context}
}
\description{
This generic returns parts of a given Facebook collections
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")
 
## Getting information about two example Facebook Pages
 fb.pages <- FacebookPagesCollection(id = c("9thcirclegames",
                                           "NathanNeverSergioBonelliEditore"), 
                                     token = fb_oauth)
 
## Pull at most 20 albums from each page
 fb.albums <- FacebookAlbumscollection(id = fb.pages, token = fb_oauth, n = 20)
 
## Create a new collection skipping the first 10 albums
 fb.oldest.albums <- fb.albums[11:length(fb.oldest.albums)]
}
}
