% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\name{as.data.frame.FacebookGenericCollection}
\alias{as.data.frame.FacebookGenericCollection}
\title{Returns a data frame from a Facebook collection}
\usage{
\method{as.data.frame}{FacebookGenericCollection}(x, row.names = FALSE,
  optional = FALSE, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}

\item{row.names}{If set to \code{TRUE}, names the rows of the returned data frame with IDs of the elements}

\item{optional}{Not used in this context.}
}
\description{
This generic return a valid data frame rapresentation of various Facebook collections
}
\keyword{internal}
