% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{as.list}
\alias{as.list}
\title{Returns a list representation of various kinds of Facebook collections}
\usage{
as.list(x, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}
}
\description{
Returns a list representation of various kinds of Facebook collections
}
\keyword{internal}

