% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.search.R
\name{facebook.search}
\alias{facebook.search}
\title{Perform a search on Facebook}
\usage{
facebook.search(query, token, n = getOption("facebook.maxitems"),
  type = "page")
}
\arguments{
\item{query}{comma-delimited string or string vector containing keywords to search.
Note that the returned results will contain any of the keywords.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{Maximum number of elements to return.}

\item{type}{string containing the type of content to get.
Valid content types are \code{user, page, event, group, place, placetopic}}
}
\description{
\code{facebook.search} retrieves public content that mention a given keyword(s) and for a given type.
Please note that each type has its unique set of fields,
as referenced on official documentation \url{https://developers.facebook.com/docs/graph-api/using-graph-api/v2.3#search}

If \code{fields} is not specified or null, the function return a data frame containing \code{id} and \code{name} for each result
plus the given \code{type}.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Searching 100 public pages that mention "facebook"
 load("fb_oauth")
	pages <- facebook.search(query="cats", type="page", token=fb_oauth, n=100)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{facebook.query}}
}

