% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookCommentsCollection.R
\name{FacebookCommentsCollection}
\alias{FacebookCommentsCollection}
\title{Build a Collection of Facebook Post Comments}
\usage{
FacebookCommentsCollection(id, token = NULL, parameters = list(),
  fields = c("id", "from.fields(id,name)", "message", "created_time",
  "like_count"), n = getOption("facebook.maxitems"), metadata = FALSE,
  .progress = create_progress_bar())
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any of the supported types}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If \code{NULL} and \code{id} is a Collection, get that one instead. Otherwise, no query is performed
to the Facebook Graph API and an empty Collection is returned}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}}

\item{fields}{A character vector or a comma-delimited string with the list of fields to get for each ID. If no value for a given fields is found, it will
be set to \code{NULL}}

\item{n}{If \code{id} is a Collection, then \code{n} is the maximum number of posts to be pulled for any element of the Collection in \code{id}.
Otherwise, the parameter is ignored. It can be set to \code{Inf} to pull out any available public post and assume the default value from the value
of \code{facebook.maxitems} global option if missing.}

\item{metadata}{If set to \code{TRUE}, the medatata from each ID is pulled with the data and the \code{type} slot is fed. Please note
that setting this to \code{TRUE} will considerably slow down the query.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console.}
}
\value{
A collection of comments in a \code{\link{FacebookCommentsCollection-class}} object.
}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook comments to posts and build a \code{FacebookPostsCollection-class}
instance.
}
\details{
\code{FacebookCommentsCollection} is the constructor for the \code{\link{FacebookCommentsCollection-class}}.
It returns data about comments but doesn't return lists of their own comments or likes (altough it could return a summary view of both).

Consider pass an instance of a Facecook Posts Collection build using the construction \code{\link{FacebookPostsCollection}},
if you need to bind a comment to its parent post.

Due to the network-graph nature of Facebook data model,
you can always specify fields details for each field eventually nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{name} for the \code{from} node, this clause is valid among others:
\code{from.fields(id,name)}.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")

## Getting information about two example Facebook Pages
 fb.pages <- FacebookPagesCollection(id = c("9thcirclegames", "NathanNeverSergioBonelliEditore"), token = fb_oauth)

## Pull the latest 10 posts from each page in a post collection
 fb.posts <- FacebookPostscollection(id = fb.pages, token = fb_oauth, n = 10)

## Pull all the available posts from each page in a post collection
 fb.posts.inf <- FacebookPostscollection(id = fb.pages, token = fb_oauth, n = Inf)

## Pull all the available comments from each post of the first collection
 fb.comments.inf <- FacebookPostscollection(id = fb.posts, token = fb_oauth)

## Convert the collection to a data frame
fb.comments <- as.data.frame(fb.comments)

# The same as before in a more compact fashion using the pipe operator
# chaining from a Pages and then a Posts Collection
fb.comments.pipe <-
 FacebookPagesCollection(id = c("9thcirclegames", "NathanNeverSergioBonelliEditore"), token = fb_oauth) \%>\%
   FacebookPostscollection(n = 10) \%>\% FacebookCommentsCollection(n = Inf)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookLikesCollection}}, \code{\link{FacebookPostsCollection}}, \code{\link{fbOAuth}}

Other Facebook Collection Costructors: \code{\link{FacebookLikesCollection}};
  \code{\link{FacebookPostsCollection}}
}

