% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookConversationsCollection.R
\name{FacebookConversationsCollection}
\alias{FacebookConversationsCollection}
\title{Build a collection of Facebook inbox conversations}
\usage{
FacebookConversationsCollection(id, token = NULL, parameters = list(),
  fields = c("id", "snippet", "updated_time", "message_count"),
  n = getOption("facebook.maxitems"), metadata = FALSE,
  .progress = create_progress_bar())
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any
of the supported types (see below).}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If token \code{NULL} and \code{id} is a collection,
use the token of the source collection. Otherwise, no query is performed
to the Facebook Graph API and an empty collection is returned.}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}.}

\item{fields}{A character vector with the fields to get for each \code{id}. If no value
for a given field is found, it will be set to \code{NULL}.}

\item{n}{If \code{id} is an iterable collection, then \code{n} is the maximum number of conversations to be pulled for each element of the source collection
in \code{id}. It can be set to \code{Inf} to pull out any available conversation and assumes the default value from the value
of \code{facebook.maxitems} global option if missing. If \code{id} is not a collection or cannot be iterated, the parameter is ignored.}

\item{metadata}{If set to \code{TRUE}, the medatata for each ID is pulled with the data
and the \code{type} slot is fed accordingly.
Please note that setting this to \code{TRUE} could considerably
slow down the execution time, as more queries are needed.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console. See \link[plyr]{create_progress_bar} for details.}
}
\value{
A collection of conversations in a \code{\link{FacebookConversationsCollection-class}} object.
}
\description{
Connect to Facebook Graph API, get information from a list of inbox conversations and build a \code{\link{FacebookConversationsCollection-class}}
instance.
}
\details{
\code{FacebookConversationsCollection} is the constructor for the \code{\link{FacebookConversationsCollection-class}}.
It returns data about inbox conversations of users or pages but doesn't return lists of messages or users involved.

A \code{\link{FacebookMessagesCollection-class}} can be built on top to pull out the messages of each conversation.
}
\section{Nesting fields}{

Due to the network-graph nature of Facebook data model,
you can specify fields details for each field nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{source} for the \code{cover} field, this is valid among others:
\code{cover.fields(id,source)}.

Following the same philosophy, if you need only \code{id} and \code{name} for the \code{from} node
you can use \code{from.fields(id,name)}.
}

\section{Valid sources}{

Instead of a character vector, one of these collections can also be passed as parameter in \code{id}:
\itemize{
 \item{\code{\link{FacebookPagesCollection-class}} will build a collection with
 all the inbox conversations to pages of the source collection. A page access token with \code{read_page_mailboxes} is needed
 to read the related mailboxes.}
 \item{\code{\link{FacebookUsersCollection-class}} will build a collection with
 all the conversations that person is involved with \strong{if they are a developer of the app making the request.}
 A user access token with \code{read_mailbox} is needed to perform this action.}
 \item{\code{\link{FacebookMixedCollection-class}} will build a collection with
 all the conversations, filtering only to the \code{user} and \code{page} in the source collection.}
}
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_page_oauth")

## Getting the conversations of an example page. A page access token is needed to access the inbox
 conversations <- FacebookConversationsCollection(FacebookPagesCollection(
                                                    "9thcirclegames", fb_page_oauth))
}
}
\author{
Gabriele Baldassarre \url{https://gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookMessagesCollection}}, \code{\link{facebook.participants}}, \code{\link{facebook.senders}}

Other Facebook Collection Costructors: \code{\link{FacebookAlbumsCollection}};
  \code{\link{FacebookCommentsCollection}};
  \code{\link{FacebookEventsCollection}};
  \code{\link{FacebookGroupsCollection}};
  \code{\link{FacebookLikesCollection}};
  \code{\link{FacebookMessagesCollection}};
  \code{\link{FacebookPagesCollection}};
  \code{\link{FacebookPhotosCollection}};
  \code{\link{FacebookPostsCollection}};
  \code{\link{FacebookUsersCollection}};
  \code{\link{FacebookVideosCollection}}
}

