% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-class.R
\docType{class}
\name{FacebookGenericCollection-class}
\alias{FacebookGenericCollection-class}
\title{A generic collection of Facebook elements}
\description{
Connect to Facebook Graph API, get public information from a list of Facebook elements of any kind and build a \code{FacebookUsersCollection-class}
instance.
}
\details{
This collection should not be built directly, as it performs no sanity check on its content.
However, among all the available collections, it's the only one that can have mixed content inside, so
many commodity endpoint functions like \code{\link{facebook.search}} fill this as return value.

If you exactly know what you're doing you \emph{could} eventually build an instance of this class to perform generic queries to Graph API,
but it's not guaranteed to work and it probably won't, actually.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector of the IDs of the item included in the Collection}

\item{\code{fields}}{A character vectorwith the content of any ID in the Collection. Ie \code{id, name, created_time}...}

\item{\code{token}}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth} used to fill the Collection.}

\item{\code{parent}}{A character vector of the parent IDs of the items included in the Collection or \code{NA} if the items have not inherited a parent}

\item{\code{parent.collection}}{The collection used to build this one, or NA if the collection was built from atomic IDs.}

\item{\code{type}}{A character vector with the type of each element of the collection. It's set only if the collection is built with \code{metadata=TRUE}}
}}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
Other Facebook Collections: \code{\link{FacebookCommentsCollection-class}};
  \code{\link{FacebookLikesCollection-class}};
  \code{\link{FacebookPostsCollection-class}},
  \code{\link{FacebookPostsCollection-class}}
}
\keyword{internal}

