% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R,
%   R/FacebookGenericCollection-methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,FacebookGenericCollection-method}
\alias{summary,ANY-method}
\title{Returns a summary of a Facebook collection}
\usage{
summary(object, ...)

\S4method{summary}{FacebookGenericCollection}(object, ...)
}
\arguments{
\item{object}{A valid collection of Facebook elements}

\item{...}{not used in this context}
}
\description{
This generic returns a summarized version from various Facebook collections
}
